Add files via upload
This is my solution for the 3MTT South West Open Source Challenge. designed a landing page for government officials to showcase their weekly activities to all Nigerians and as well for Nigerians to make a request to any government officials on pressing matters that need the government attention. Please kindly find attached my Authentication and Landing page including the Design system and User flow